requirements.txt
streamlit~=1.26.0
torch~=2.0.1
torchvision~=0.15.2
torchaudio
accelerate~=0.21.0
diffusers[torch]~=0.15.0.dev0
transformers~=4.31.0
Pillow~=9.5.0
PyYAML~=6.0.1
tqdm~=4.66.1